Moved configuration structs into own files
- PVPConfiguration
- RCONConfiguration
- CompressionConfiguration
- CommandsConfiguration
